Implement local web UI for development
